+++ src/store/store.tsx
@@ -2,7 +2,7 @@ import { combineReducers, configureStore } from "@reduxjs/toolkit";
 import imageSearchReducer from "./imageSearch/imageSearch.slice";
 import formDataReducer from "./formData/formData.slice";
 import employeesReducer from "./employees/employeesSlice";
-import getImagesReducer, { getImages } from "./getImages/getImages.slice";
+import { getImages } from "./getImages/getImages.slice";
 import { TypedUseSelectorHook, useSelector } from "react-redux";
 import { setupListeners } from "@reduxjs/toolkit/query";
 
@@ -10,13 +10,11 @@ export const reducers = combineReducers({
   search: imageSearchReducer,
   form: formDataReducer,
   employees: employeesReducer,
-  // getImages: getImagesReducer,
   [getImages.reducerPath]: getImages.reducer,
 });
 export const store = configureStore({
   reducer: reducers,
-  // middleware: (getDefaultMiddleware) =>
-  //   getDefaultMiddleware().concat(getImages.middleware),
+
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({
       serializableCheck: false,
