+++ src/types/pixabayAPI.tsx
@@ -1,0 +1,35 @@
+export interface IImageResponseItem {
+  collections: number;
+  comments: number;
+  downloads: number;
+  id: number;
+  imageHeight: number;
+  imageSize: number;
+  imageWidth: number;
+  largeImageURL: string;
+  likes: number;
+  pageURL: string;
+  previewHeight: number;
+  previewURL: string;
+  previewWidth: number;
+  tags: string;
+  type: string;
+  user: string;
+  userImageURL: string;
+  user_id: number;
+  views: number;
+  webformatHeight: number;
+  webformatURL: string;
+  webformatWidth: number;
+}
+
+export interface IImageResponseResult {
+  hits: IImageResponseItem[];
+  total: number;
+  totalHits: number;
+}
+
+export interface IQueryParams {
+  query: string;
+  page: number;
+}
